requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
groq>=0.9
python-dotenv>=1.0
pydantic>=2.6
Pillow>=10.3
cloudinary>=1.40
pytest>=8.0
pytest-asyncio>=0.23
